- Refactored StructuralLevel out from the items package since it's used in so many places, while at it, pulled all client code from it into a client only class - Did some minor general cleanup on the various classes while I was in them, mostly indentation -> tabs, cleaning up imports, curly brace formatting, removing some unneeded casts, adding annotations - In client data gen(textures/translation) for a structural tier, if the base icon is not able to be loaded for a given OreDict entry, it will retry all following entries until an icon is able to be loaded. If none load, it will use black as a base color

+++ src/main/java/mcp/mobius/betterbarrels/common/blocks/TileEntityBarrel.java
@@ -653,7 +653,7 @@ public class TileEntityBarrel extends TileEntity implements ISidedInventory, IDe
 	@Override
 	public int getInventoryStackLimit() {return 64;}
 	@Override
-	public boolean isUseableByPlayer(EntityPlayer var1) {return this.worldObj.getTileEntity(this.xCoord, this.yCoord, this.zCoord) != this ? false : var1.getDistanceSq((double)this.xCoord + 0.5D, (double)this.yCoord + 0.5D, (double)this.zCoord + 0.5D) <= 64.0D;}
+	public boolean isUseableByPlayer(EntityPlayer var1) {return this.worldObj.getTileEntity(this.xCoord, this.yCoord, this.zCoord) != this ? false : var1.getDistanceSq(this.xCoord + 0.5D, this.yCoord + 0.5D, this.zCoord + 0.5D) <= 64.0D;}
 	@Override
 	public void openInventory() {}
 	@Override
